protege_backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
groq>=0.4.0
firebase-admin>=6.3.0
httpx>=0.26.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
